refactor: Fix misc
