+++ cachez/errors.go
@@ -1,0 +1,5 @@
+package cachez
+
+import "errors"
+
+var ErrAssertionFailed = errors.New("assertion failed")
