+++ cachez/lru.go
@@ -171,10 +171,3 @@ func (c *lruCacheG[K, V]) removeElement(elem *list.Element) {
 
 	delete(c.items, ent.key)
 }
-
-//nolint:gochecknoglobals,revive
-var (
-	_testAssert_lruCacheG_Get           bool
-	_testAssert_lruCacheG_SetWithTTL    bool
-	_testAssert_lruCacheG_removeElement bool
-)
